Time: 0 ms (100.00%), Space: 12.6 MB (72.11%) - LeetHub

+++ 0260-single-number-iii/0260-single-number-iii.cpp
@@ -6,7 +6,7 @@ public:
         for(int i = 0; i <n; i++) {
             xorr = xorr ^ nums[i];
         }
-        long rightMost = (xorr & (xorr - 1)) ^ xorr;
+        int rightMost = (xorr & (xorr - 1)) ^ xorr;
         int b1 = 0;
         int b2 = 0;
